Added controllers to follow the MVC method, routes are clean now
