+++ routes/users.js
@@ -1,10 +1,9 @@
 var express = require('express');
 var router = express.Router();
 const mongoose = require('mongoose');
-const User = require('../task-manager/src/models/user.js');
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
+
 const checkAuth = require('../task-manager/src/middleware/check-auth');
+const UsersController = require('../task-manager/src/controller/users');
 
 /**Put you DB path here, you can use this default path to host it local at this address */
 mongoose.connect('mongodb://localhost/test', {
@@ -13,115 +12,17 @@ mongoose.connect('mongodb://localhost/test', {
   useCreateIndex: true
 });
 
-/*POST users listing, posting to /users with json will create entry in DB */
-/***Currently hashes password using bcrypt, it also checks if email was used and wont let another user be created with the same email twice */
-router.post('/signup', (req, res, next) => {
-  User.find({ email: req.body.email })
-    .exec()
-    .then((user) => {
-      if (user.length >= 1) {
-        return res.status(409).json({
-          message: 'Email exists'
-        });
-      } else {
-        bcrypt.hash(req.body.password, 10, (err, hash) => {
-          if (err) {
-            return res.status(500).json({
-              error: err
-            });
-          } else {
-            const user = new User({
-              displayName: req.body.displayName,
-              email: req.body.email,
-              password: hash
-            });
-            user
-              .save()
-              .then((result) => {
-                console.log(result);
-                res.status(201).json({
-                  message: 'User Created',
-                  createdUser: user
-                });
-              })
-              .catch((err) => {
-                console.log(err);
-                res.status(500).json({
-                  error: err
-                });
-              });
-          }
-        });
-      }
-    })
-    .catch();
-});
+/*POST user signup, posting to /users/signup with json will create entry in DB */
+router.post('/signup', UsersController.user_sign_up);
 /**************************************************************************/
 
 /*POST to login exsiting user************************************************/
-/*Uses jwt to prodeuce web token************************************************/
-router.post('/login', (req, res, next) => {
-  User.find({ email: req.body.email })
-    .exec()
-    .then((user) => {
-      if (user.length < 1) {
-        return res.status(401).json({
-          message: 'Auth failed'
-        });
-      }
-      bcrypt.compare(req.body.password, user[0].password, (err, result) => {
-        if (err) {
-          return res.status(401).json({
-            message: 'Auth failed'
-          });
-        } else if (result) {
-          const token = jwt.sign(
-            {
-              email: user[0].email,
-              userId: user[0]._id
-            },
-            process.env.JWT_KEY,
-            {
-              expiresIn: '1h'
-            }
-          );
-          return res.status(200).json({
-            message: 'Auth successful',
-            token: token
-          });
-        }
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.post('/login', UsersController.user_login);
 
 /**************************************************************************/
 
 /*GET all displayNames! ***********************************************************/
-router.get('/', (req, res, next) => {
-  User.find()
-    .select('displayName')
-    .exec()
-    .then((docs) => {
-      const response = {
-        count: docs.length,
-        users: docs
-      };
-      console.log(response);
-      res.status(200).json(docs);
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.get('/', UsersController.user_display_names);
 /**************************************************************************/
 
 module.exports = router;
